--- src/ILayoutBuilder.py
@@ -1,8 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class ILayoutBuilder(ABC):
-
-    @abstractmethod
-    def build(self):
-        pass
